lec_6_files/rag/requirements.txt
Flask
streamlit
pandas
numpy
python-dotenv
langchain-community
langchain-openai
langchain